% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhs_api_url.R
\docType{data}
\name{dhs_api_url}
\alias{dhs_api_url}
\title{DHS API URL Base}
\format{A character vector (length 1)}
\usage{
dhs_api_url
}
\description{
Just a character vector to use for the base URL for the
DHS API
}
\keyword{datasets}
