% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhs_countries.R
\name{dhs_countries}
\alias{dhs_countries}
\title{Return DHS Countries}
\usage{
dhs_countries()
}
\value{
A \code{data.frame} of the listing of countries
}
\description{
Returns the set of countries from DHS from their website
}
\examples{
head(dhs_countries())
}
