% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhs_indicators.R
\name{dhs_indicators}
\alias{dhs_indicators}
\title{Return DHS Indicators}
\usage{
dhs_indicators()
}
\value{
A \code{data.frame} of the listing of indicators
}
\description{
Returns the set of indicators from DHS from their website
}
\examples{
head(list_of_indicators())
}
