% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhs_survey_characteristics.R
\name{dhs_survey_characteristics}
\alias{dhs_survey_characteristics}
\title{Return DHS Survey Characteristics}
\usage{
dhs_survey_characteristics()
}
\value{
A \code{data.frame} of the listing of Survey
Characteristics
}
\description{
Returns the set of Survey Characteristics from
DHS from their website
}
\examples{
head(dhs_survey_characteristics())
}
