% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhs_surveys.R
\name{dhs_surveys}
\alias{dhs_surveys}
\title{Return DHS Surveys}
\usage{
dhs_surveys()
}
\value{
A \code{data.frame} of the listing of surveys
}
\description{
Returns the set of surveys from DHS from their website
}
\examples{
head(dhs_surveys())
}
