% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhs_tags.R
\name{dhs_tags}
\alias{dhs_tags}
\title{Return DHS Tags}
\usage{
dhs_tags()
}
\value{
A \code{data.frame} of the listing of tags
}
\description{
Returns the set of tags from DHS from their website
}
\examples{
head(dhs_tags())
}
