% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dhs_data.R
\name{get_dhs_data}
\alias{get_dhs_data}
\title{Get DHS data from a URL}
\usage{
get_dhs_data(url, query = NULL, progress = TRUE)
}
\arguments{
\item{url}{URL for DHS}

\item{query}{Query to pass to \code{\link{GET}}}

\item{progress}{Show progress bar if more than one page}
}
\value{
List of the data, the number of pages parsed, the number of
records to be in the output, the number of records in the output.
}
\description{
Worker function for running and returning DHS data
}
