% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fields.R
\name{get_fields}
\alias{get_fields}
\title{Get DHS Fields}
\usage{
get_fields()
}
\value{
\code{data.frame} of the field names and the description of
the field
}
\description{
Returns a \code{data.frame} of the fields from
\url{http://api.dhsprogram.com/rest/dhs/data/fields}
}
\examples{
get_fields()
}
