% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_data.R
\name{indicator_data}
\alias{indicator_data}
\title{Get DHS Fields}
\usage{
indicator_data(per_page = 1000, progress = TRUE, survey = NULL,
  countries = NULL, indicators = NULL, years = NULL)
}
\arguments{
\item{per_page}{Number of records to return per page}

\item{progress}{Show progress bar if more than one page}

\item{survey}{Surveys to return based on survey ID}

\item{countries}{Countries to return based on country ID}

\item{indicators}{Indicators to return based on country ID}

\item{years}{Years to return}
}
\value{
\code{data.frame} of the field names and the description of
the field
}
\description{
Returns a \code{data.frame} of the fields from
\url{http://api.dhsprogram.com/rest/dhs/data/fields}
}
\examples{
indicator_data(survey = "SN2010DHS")
indicator_data(survey = "EG1988DHS", countries = "EG")

}
