% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind_dhs_data.R
\name{rbind_dhs_data}
\alias{rbind_dhs_data}
\title{Coerce DHS data into a \code{data.frame}}
\usage{
rbind_dhs_data(data)
}
\arguments{
\item{data}{Slot of \code{data} in output
from \code{\link{get_dhs_data}}}
}
\value{
A \code{data.frame} of the data.  Most likely all columns are
going to be character.
}
\description{
Worker function for running and returning DHS data,
but in a \code{data.frame}.
}
