% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_do_var_labels.R
\name{read_do_var_labels}
\alias{read_do_var_labels}
\title{Read DO file Variable Label Definitions}
\usage{
read_do_var_labels(file)
}
\arguments{
\item{file}{.do file of label variable definitions from stata}
}
\value{
A \code{data.frame} of variable labels
}
\description{
Read DO file Variable Label Definitions
}
