% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest_dhs_url.R
\docType{data}
\name{rest_dhs_url}
\alias{rest_dhs_url}
\title{DHS URL for RESTful API}
\format{A character vector (length 1)}
\usage{
rest_dhs_url
}
\description{
Just a character vector to use for the DHS RESTful API
}
\keyword{datasets}
